+++ shell/SUDO_USER.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+# if oridinary user run sudo, then SUDO_USER env will be set.
+echo SUDO_USER=$SUDO_USER